add env configs and fix service worker (#14)
Now using the command `yarn use-config ...` we can switch between configs
locally. During deploys the appropriate env config will be copied into
place.

Additionally this fixes the service worker so that it doesn't fetch all
the app files twice. Instead it first loads itself then begins pre-caching
all app files, 10 at a time. This results in a slightly slower startup
time but saves on bandwidth and ensures the user doesn't accidentally
exit the page thinking that offline mode is ready when it isn't.

This fix required some package patching, which live in the `patches`
folder and utilize `patch-package`.

Finally this fixes the Azure Static Web App config so it properly
rewrites requests and allows any route to be accessed on first load by
going directly to it. e.g. /passages/100100