add eslint plugin to lint for unused/missing translations (#43)
It's really easy to mess up translations since they're strings that map between a usage and a language file. ESLint is the perfect tool to make sure we don't end up committing translation files that are missing strings, or translation files that have unused strings.

The approach is to check for missing translations in the first pass ESLint does, saving all
translation strings to a tmp file. Then a second pass is done with only the unused translations rule
to see if the translation files have any strings that don't get referenced anywhere.

This also fixes the use-config so it works with .env.global.