updates for Hindi content to work (#55)
This adds Hindi to the languages list on the main page and Hindi I18N translations. It also allows audioTimestamps to be missing for books of the Bible (since for the Indian Revised Version (IRV) we don't have timestamps for the NT). Finally it adds a loading indicator to the Book on the main page while the content is loading.