add resource pane to passage page
The resource pane should come up from the bottom of the screen. It's available on mobile and
desktop, though it looks nicer on mobile. There's also some new logic included to hide irrelevant
buttons on desktop. After selecting an image, the resource pane should dismiss. It can also be
dismissed by clicking elsewhere on the screen, tapping the Resources button again, or sliding down
(on mobile).