add fallback pattern for selecting Bible if content not available in current language (#67)
When viewing a passage that is not available in the current language, show a page with options to
view the passage in a different translation.

The way it works is a query param bibleLanguage is set in the URL with whatever language the user
selects. This means that the back button will still work if they go to another page and also allows refreshes
without losing the context.

To enable some of what's happening I updated the way that languageId and languageCode can be used to lookup language info.