add recording functionality for passages (#69)
When viewing a passage, there's now a gear icon at the top right of the nav bar. It will open a modal that allows the user to record on their device. They can set the language, version, and versionAbbreviation, then save.

It will save the recording to the Cache and then use localForage to save the metadata to IndexedDb.

Additionally, this PR puts in place a pattern for feature flags. New flags can be added to the feature-flags.store.ts and can be used as any other store like $featureFlags.audioRecording. To flip a feature flag on or off, on Desktop put #ff at the end of the URL you're on and it will pop up a dialog allowing you to turn any flag on or off. The flag will be saved to localStorage.

On mobile, tap the following sequence: Top Left, Top Right, Top Right, Bottom Left and it will pop up the modal.