make audio player work with multiple files (#78)
The AudioPlayer now has two options:

OPTION 1: Pass in a list of files you want to play. Great for when you just need one AudioPlayer to play one or a sequence of clips.
OPTION 2: Pass a mapping of MultiClipAudioState, indexed by string keys. Great for when you want one AudioPlayer on a page that can quickly switch between playing different sequences of clips.
The "clip sequence" functionality is used for Bible audio, so that a passage spanning multiple chapters can play both chapter clips back to back, as if it were one clip.

The multi-clip state functionality is used by the passage page as a whole so that one AudioPlayer can be rendered at the bottom and the state swapped between based on what is showing.