update backend calls on home and passage consume pages (#87)
This swaps existing backend calls for the new API calls. There is a lot of duplication in code as it relates to the file manager, but I avoided touching file manager stuff as much as possible to not mess with @carlosmorales125's work.

One thing that is still needed is the CBBT-ER audio handling, since we'll need to unzip the combined audio files before loading them into the audio players.