update audio player to work in iOS and Safari
One big thing wrong was the cache wasn't set right for the CDN so the range requests Safari sends
were getting cached and then blowing up when accessing the cache later.

Also when unzipping the files, we needed to make sure the Blob was getting set with the right MIME
type.

Lastly this also fixes some small bugs that were present in the audio player even on Chrome, mostly
around playing multiple clips.