switch to custom service worker with custom plugins
This moves away from the generated service worker toward a custom one that gets injected with the
manifest values. It does this to enable:
- a custom strategy that uses CacheFirst until an expiration time, when it switches to
  StaleWhileRevalidate
- a custom plugin to handle caching "partial" responses due to range requests with audio/video files

Additionally, it adds a modal that asks the user if they want to update when changes are detected
rather than assuming they want to and refreshing the page.