switch to custom service worker with custom plugins (#95)
This moves away from the generated service worker toward a custom one that gets injected with the
manifest values. It does this to enable:

- a custom strategy that uses CacheFirst until an expiration time, when it switches to StaleWhileRevalidate
- a custom plugin to handle caching "partial" responses due to range requests with audio/video files

Additionally, it removes the auto-refresh when a service worker update is detected. This was causing the app to refresh at weird times, sometimes canceling user actions (like selecting from the home page dropdowns). Instead, we'll wait for the user to refresh or for the user to close and reopen the site.