improve CBBT-ER step button carousel and fix dropdown issues (#117)
In order to make it easier for people to switch steps on desktop (and mobile to a lesser extent),
this improves the auto-scroll behavior of the button carousel. Now it will always scroll into view
the button to the right of the current step.

Also addresses some small things related to dropdowns.