handle multiple Bibles on the passage page
On the passage page there are now tabs at the top for different Bibles (similar to CBBT-ER steps).

When viewing a passage online, the new PreferredBiblesModal will allow selecting which Bibles a user
wants to have tabs for.

When viewing a passage offline, all Bibles that are downloaded for that passage will be shown.