improve error logging util to handle network errors more smartly (#137)
This updates the logging util to not send to app insights on network errors. It also adds some missing code that hooks into unhandled rejections app-wide to make sure all errors are being handled. Previously only errors in page.ts and on navigations would be handled, which we weren't really using since we're in SPA mode.

Additionally, existing try/catch blocks have been updated to make sure we're logging in every case. This will be fine since now only non-network errors will be sent to app insights.

Finally, there's a new dev-local environment that's meant to be used locally that does not include the app insights key so we're not cluttering logs.