fix undefined error and add stricter TS settings to catch future issues (#162)
We were missing some optional chaining code in a spot that was causing silent errors for end users when loading content in a language without a default Bible. Note it was just a silent error and the user wouldn't see anything wrong, but it was an error nonetheless.

After fixing I realized we probably want TS to be stricter so it catches these before users hit them. I added a config option that is stricter around this issue: noUncheckedIndexedAccess