refactor guide to its own component in prep for displaying uWTranslationNotes (#223)
In order to display unfoldingWord Translation Notes as its own guide type, we need to split off the FIA-specific things into its own component. This is done by creating a GuideContent component that decides what to render based on the selected guid type. The only thing it renders right now is FiaContent.

In addition, this renames CBBT-ER to FIA for consistency.