batch fetches for unfoldingWord Translation Notes for faster loads (#231)
Currently we use the batch endpoints for the download manager but not for pulling content for display.

This PR utilizes the batch endpoints for displaying unfoldingWord Translation Notes content. Rather than needing to fetch each content and metadata individually for each note (which there are about 2-5 per verse), we can batch all of the loads together.

However it's not as simple as just calling the batch endpoint every time since we aren't actually caching the batch endpoint URLs themselves. Instead we use the batch endpoint response to manually cache the individual content data.

This causes two problems:

When the user is offline if we tried a batch endpoint call it would fail.
When the user is online we will be fetching data that the user may already have cached. It essentially means every time we need to hit the server since it won't see anything as cached.
The solution is to first split the items based on if they're cached yet or not. Then call the batch endpoint for the uncached items and just do regular fetches for the cached items (since it'll just pull from cache).