refactor caching to central config (#238)
This is a large refactor of the caching logic and renamed references to CDN to better fit with how things currently work.

Before we had an API and a CDN cache and the divide was very explicit. However now that most content lives in the API (except for media) that distinction doesn't make sense. I've renamed it to API and Content caches to make it clear that any content is meant for the content cache (even if it comes from the API), and API holds the rest.

Along with making that distinction clearer this also moves all URL regex handling into one file that can be used in the service worker and the app to determine which cache a given URL belongs to.