Treat verse 0 as superscript. (#384)
Many Psalms have superscript data. Right now the superscript is included
in the verse 1 text for all Bibles. We are moving toward extracting out
the superscripts and considering them verse "0" which allows for better
versification mappings between Bible translations. It also will allow us
to easily display the superscript differently. This PR is backwards
compatible if there is no verse 0 data for a Psalm (see first
screenshot) but will render verse 0 as an italicized superscript if it
is present (see second screenshot). The verse picker still only shows
verse 1 as the first verse in the Psalm.

Audio data for the superscript is still associated with verse 1. Thus,
if you choose only verse 1 it will render under the hood both verse 0
and verse 1 and display them both as verse 1, which is still correct for
audio timings.

Hebrew alignments are not yet supported so this doesn't need to handle
alignment word mappings. If we do eventually add OT alignment data, it
uses verse 0 for superscripts so this should be forwards compatible.

Note that if you do manually type verse 0 into the URL everything will
still work correctly.

No superscript:

![image](https://github.com/user-attachments/assets/ada3f036-d942-4d36-b4d6-5059047ddb3f)

Superscript:

![image](https://github.com/user-attachments/assets/a1dfecf5-b754-4575-a804-0b50353eafa2)